requirements.txt

openai==2.8.0
pillow==12.0.0
requests==2.32.5


